backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
earthengine-api==1.4.1
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
pytest==8.3.3
httpx==0.27.2
